Valkey-cli:  add missing -c to a test to fix resharding failures
A test was missing the `-c` flag for valkey-cli, which caused it to
miscount the number of databases and led to resharding failures. This
change adds the flag to fix the test. "-c" enables cluster mode
in the cli.

Note: valkey-cli will be updated separately to not rely on `-c` to
determine the number of databases.

Signed-off-by: xbasel <103044017+xbasel@users.noreply.github.com>