replace vset "virtual table" with explicit function parameters to get entry expiry.
This is the first change in order to reduce the vset default memory consumption.
Although this complicates the API, it allows reducing the memory footprint per
each hash object using the set.

The next potential step is to make the vset a pure bucket pointer so that it will not use any extra memory.

I intentionally separated these changes in order for us to be able to decide if "sacrifice" API friendly is better
than consuming more memory

Signed-off-by: Ran Shidlansik <ranshid@amazon.com>